third_party/WebKit/LayoutTests/fast/forms/select/menulist-popup-item-style-expected.txt

Tests fix for regression crbug.com/395613. Checks that option in optgroup is styled properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isSelectPopupVisible(popup) is true
PASS internals.selectPopupItemStyleIsRtl(popup, 2) is true
PASS successfullyParsed is true

TEST COMPLETE
